requirements.txt
certifi==2021.10.8
cycler==0.10.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
Pillow==8.4.0
pygame==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
six==1.16.0
wincertstore==0.2
torch==1.9.1
cloudpickle==1.6.0
matplotlib==3.3.4
opencv-python==4.5.3.56
scikit-learn==0.24.2
scipy==1.6.0
tqdm==4.56.0
pyyaml
tensorboard
gym
tabulate

